# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_iLLeniumStudios_cronjob-guardian_api_v1alpha1.AlertChannel:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/github_com_iLLeniumStudios_cronjob-guardian_api_v1alpha1.AlertChannelSpec'
      status:
        $ref: '#/definitions/github_com_iLLeniumStudios_cronjob-guardian_api_v1alpha1.AlertChannelStatus'
    type: object
...
  internal_api.ActiveJobItem:
    properties:
      name:
        type: string
      podName:
        type: string
      podPhase:
        type: string
      ready:
        type: string
      runningDuration:
        type: string
      startTime:
        type: string
    type: object
...
  internal_api.AlertHistoryItem:
    properties:
      channelsNotified:
        items:
          type: string
        type: array
      cronjob:
        $ref: '#/definitions/internal_api.NamespacedRef'
      exitCode:
        description: Context fields for failure alerts
        type: integer
      id:
        type: string
      message:
        type: string
      occurredAt:
        type: string
      reason:
        type: string
      resolvedAt:
        type: string
      severity:
        type: string
      suggestedFix:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  internal_api.AlertHistoryResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api.AlertHistoryItem'
        type: array
      pagination:
        $ref: '#/definitions/internal_api.Pagination'
    type: object
  internal_api.AlertItem:
    properties:
      context:
        $ref: '#/definitions/internal_api.AlertContextResponse'
      cronjob:
        $ref: '#/definitions/internal_api.NamespacedRef'
      id:
        type: string
      lastNotified:
        type: string
      message:
        type: string
      monitor:
        $ref: '#/definitions/internal_api.NamespacedRef'
      severity:
        type: string
      since:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  internal_api.AlertListResponse:
    properties:
      bySeverity:
        additionalProperties:
          type: integer
        type: object
      items:
        items:
          $ref: '#/definitions/internal_api.AlertItem'
        type: array
      total:
        type: integer
    type: object
  internal_api.ChannelListItem:
    properties:
      config:
        additionalProperties: {}
        type: object
      lastTest:
        $ref: '#/definitions/internal_api.TestResult'
      name:
        type: string
      ready:
        type: boolean
      stats:
        $ref: '#/definitions/internal_api.ChannelStats'
      type:
        type: string
    type: object
  internal_api.ChannelListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api.ChannelListItem'
        type: array
      summary:
        $ref: '#/definitions/internal_api.ChannelSummary'
    type: object
...
  internal_api.ChannelSummary:
    properties:
      notReady:
        type: integer
      ready:
        type: integer
      total:
        type: integer
    type: object
...
  internal_api.CronJobDetailResponse:
    properties:
      activeAlerts:
        items:
          $ref: '#/definitions/internal_api.AlertItem'
        type: array
      activeJobs:
        items:
          $ref: '#/definitions/internal_api.ActiveJobItem'
        type: array
      lastExecution:
        $ref: '#/definitions/internal_api.ExecutionSummary'
      metrics:
        $ref: '#/definitions/internal_api.CronJobMetrics'
      monitorRef:
        $ref: '#/definitions/internal_api.NamespacedRef'
      name:
        type: string
      namespace:
        type: string
      nextRun:
        type: string
      schedule:
        type: string
      status:
        type: string
      suspended:
        type: boolean
      timezone:
        type: string
    type: object
...
  internal_api.CronJobListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api.CronJobListItem'
        type: array
      summary:
        $ref: '#/definitions/internal_api.SummaryStats'
    type: object
...
  internal_api.DeleteHistoryResponse:
    properties:
      message:
        type: string
      recordsDeleted:
        type: integer
      success:
        type: boolean
    type: object
  internal_api.ErrorDetail:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  internal_api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_api.ErrorDetail'
    type: object
  internal_api.ExecutionDetailResponse:
    properties:
      completionTime:
        type: string
      cronJobName:
        type: string
      cronJobNamespace:
        type: string
      cronJobUID:
        type: string
      duration:
        type: string
      exitCode:
        type: integer
      id:
        type: integer
      isRetry:
        type: boolean
      jobName:
        type: string
      reason:
        type: string
      retryOf:
        type: string
      startTime:
        type: string
      status:
        type: string
      storedEvents:
        type: string
      storedLogs:
        type: string
    type: object
...
  internal_api.ExecutionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api.ExecutionItem'
        type: array
      pagination:
        $ref: '#/definitions/internal_api.Pagination'
    type: object
...
  internal_api.LogsResponse:
    properties:
      container:
        type: string
      jobName:
        type: string
      logs:
        type: string
      truncated:
        type: boolean
    type: object
...
  internal_api.MonitorListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api.MonitorListItem'
        type: array
    type: object
...
  internal_api.Pagination:
    properties:
      hasMore:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_api.PatternTestRequest:
    properties:
      pattern:
        $ref: '#/definitions/internal_api.PatternInput'
      testData:
        $ref: '#/definitions/internal_api.PatternTestData'
    type: object
  internal_api.PatternTestResponse:
    properties:
      error:
        type: string
      matched:
        type: boolean
      renderedSuggestion:
        type: string
    type: object
  internal_api.PruneRequest:
    properties:
      dryRun:
        type: boolean
      olderThanDays:
        type: integer
      pruneLogsOnly:
        type: boolean
    type: object
  internal_api.PruneResponse:
    properties:
      cutoff:
        type: string
      dryRun:
        type: boolean
      message:
        type: string
      olderThanDays:
        type: integer
      recordsPruned:
        type: integer
      success:
        type: boolean
    type: object
  internal_api.SimpleResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_api.StatsResponse:
    properties:
      activeAlerts:
        type: integer
      executionsRecorded24h:
        type: integer
      summary:
        $ref: '#/definitions/internal_api.SummaryStats'
      totalCronJobs:
        type: integer
      totalMonitors:
        type: integer
    type: object
  internal_api.StorageStatsResponse:
    properties:
      executionCount:
        type: integer
      healthy:
        type: boolean
      logStorageEnabled:
        type: boolean
      retentionDays:
        type: integer
      storageType:
        type: string
    type: object
  internal_api.SummaryStats:
    properties:
      critical:
        type: integer
      healthy:
        type: integer
      running:
        type: integer
      suspended:
        type: integer
      warning:
        type: integer
    type: object
...
  internal_api.TriggerResponse:
    properties:
      jobName:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: CronJob Guardian
    url: https://github.com/iLLeniumStudios/cronjob-guardian
  description: Kubernetes CronJob monitoring and alerting API. Provides SLA tracking,
    dead-man switches, and intelligent alerting for CronJobs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CronJob Guardian API
  version: "1.0"
paths:
  /admin/prune:
    post:
      consumes:
      - application/json
      description: Manually triggers pruning of old execution records
      parameters:
      - description: Prune options
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api.PruneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.PruneResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Trigger history pruning
      tags:
      - Admin
  /admin/storage-stats:
    get:
      description: Returns storage backend statistics and health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.StorageStatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get storage statistics
      tags:
      - Admin
  /alerts:
    get:
      description: Returns all active alerts across all monitored CronJobs
      parameters:
      - description: Filter by severity (critical, warning)
        in: query
        name: severity
        type: string
      - description: Filter by alert type
        in: query
        name: type
        type: string
      - description: Filter by CronJob namespace
        in: query
        name: namespace
        type: string
      - description: Filter by CronJob name
        in: query
        name: cronjob
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AlertListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List active alerts
      tags:
      - Alerts
  /alerts/history:
    get:
      description: Returns paginated history of past alerts from the store
      parameters:
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Page offset
        in: query
        name: offset
        type: integer
      - description: Filter by severity
        in: query
        name: severity
        type: string
      - description: Filter since timestamp (RFC3339)
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AlertHistoryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get alert history
      tags:
      - Alerts
  /channels:
    get:
      description: Returns all configured alert channels with their status and stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.ChannelListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List alert channels
      tags:
      - Channels
  /channels/{name}:
    get:
      description: Returns details for a specific alert channel
      parameters:
      - description: Channel name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_iLLeniumStudios_cronjob-guardian_api_v1alpha1.AlertChannel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get channel details
      tags:
      - Channels
  /channels/{name}/test:
    post:
      description: Sends a test alert to verify channel configuration
      parameters:
      - description: Channel name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimpleResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Test alert channel
      tags:
      - Channels
...
  /cronjobs:
    get:
      description: Returns all monitored CronJobs with their status
      parameters:
      - description: Filter by namespace
        in: query
        name: namespace
        type: string
      - description: Filter by status (healthy, warning, critical)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.CronJobListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List CronJobs
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}:
    get:
      description: Returns detailed information about a specific CronJob
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.CronJobDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get CronJob details
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/executions:
    get:
      description: Returns paginated execution history for a CronJob
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Page offset
        in: query
        name: offset
        type: integer
      - description: Filter by status (success, failed)
        in: query
        name: status
        type: string
      - description: Filter since timestamp (RFC3339)
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.ExecutionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get execution history
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/executions/{jobName}:
    get:
      description: Returns full execution details including stored logs and events
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      - description: Job name (execution ID)
        in: path
        name: jobName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.ExecutionDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get execution details with logs
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/executions/{jobName}/logs:
    get:
      description: Returns container logs from a job execution
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      - description: Job name
        in: path
        name: jobName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LogsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get execution logs
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/history:
    delete:
      description: Deletes all execution history for a specific CronJob
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.DeleteHistoryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Delete CronJob history
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/resume:
    post:
      description: Resumes a suspended CronJob to allow scheduled runs
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Resume CronJob
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/suspend:
    post:
      description: Suspends a CronJob to prevent scheduled runs
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Suspend CronJob
      tags:
      - CronJobs
  /cronjobs/{namespace}/{name}/trigger:
    post:
      description: Manually triggers a CronJob to run immediately
      parameters:
      - description: CronJob namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.TriggerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Trigger CronJob
      tags:
      - CronJobs
...
  /monitors:
    get:
      description: Returns all CronJobMonitor resources
      parameters:
      - description: Filter by namespace
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MonitorListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List monitors
      tags:
      - Monitors
  /monitors/{namespace}/{name}:
    get:
      description: Returns detailed information about a specific CronJobMonitor
      parameters:
      - description: Monitor namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Monitor name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Monitor details (CRD format)
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get monitor details
      tags:
      - Monitors
  /patterns/test:
    post:
      consumes:
      - application/json
      description: Tests a suggested fix pattern against sample data to verify matching
      parameters:
      - description: Pattern and test data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.PatternTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.PatternTestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Test suggested fix pattern
      tags:
      - Patterns
  /stats:
    get:
      description: Returns aggregate statistics for all monitored CronJobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.StatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get statistics
      tags:
      - System
schemes:
- http
- https
swagger: "2.0"